pyproject.toml
[project]
name = "aichatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.1",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
